FIX: @import() now supports full & relative paths
This patch fixes the issue with LessCSS not working with absolute paths, such as:

  @import("/home/someuser/less/import-file");
  @import("~/less/import-file");

The advantage of using full paths, is that you can have a single central repository of CSS/Less files with commonly used components and include those at will in your current project.

Also added is raising an ImportError when given a HTTP url in @import.

  @import url("http://lesscss.org/all.css");


GOTCHAS:

There are still some issues with using variables and mixins from imported files that are not in the same location, ie:

  #/central repository
  /home/someuser/less/global.less

  #/project repository
  /home/someuser/projects/myproject/less/mixins.less

In this case, neither files can depend securely on each other.

I added a pending spec on this issue, as I was not entirely sure how to test for it correctly.